move fasttrack to pages/
